refactor(electron): extract embedding queries out of db/queries.ts
electron/db/queries.ts had grown to 1954 lines. Move the note + task
embedding queries (note_embeddings + task_embeddings: upsert/get/delete,
projections, CardStub fetch, the Record/Row types + row mappers) into
electron/db/embeddings-queries.ts (~300 lines), depending only on the Database
handle and ts() from ./utils. queries.ts re-exports the module (export * from
"./embeddings-queries") so every ../db/queries / ./queries import
(embeddings/service.ts, db/graph-queries.ts, mcp/tools/*) is unchanged.
queries.ts drops 1954 -> 1671. Mirrors the mobile split (1c086a0).

Same governance preserved: no Database is constructed in the new file (runs on
the caller's handle), so it stays safe to import from both the Electron main
process and the esbuild-bundled MCP server.

Verified: type-check:all clean; eslint clean; electron/db/queries.test.ts (51)
+ full suite (1128) green; all four esbuild bundles (main, mcp-server,
embeddings-server, runtime-server) compile.
