+++ changelogs/v2.4.14.md
@@ -4,3 +4,7 @@
 
 - **See exactly what differs in a sync conflict.** When the same note was edited on two devices at once, the conflict-resolution dialog used to show the current note and the conflicted copy as two plain blocks of text side by side — leaving you to spot the difference by eye. It now shows a unified diff: lines only in the conflicted copy are highlighted green (+), lines only in your current note are highlighted red (−), and unchanged lines are dimmed, with a summary of how many lines changed. If the two bodies are actually identical (the conflict was in metadata only) it says so instead of showing an empty diff. Applies to both desktop and mobile.
 - **Multi-word search now finds what you mean.** Search (⌘K, the notes filter box, and the AI assistant's note/task search) previously required your whole query to appear as one exact phrase — so "auth flow" would miss a note titled "Authentication flow" or one saying "flow for auth". It now matches when every word in your query appears somewhere in the title or body, in any order. Single-word and exact-phrase searches are unchanged.
+
+### Under the hood
+
+- Split the note + task embedding queries out of the 1,950-line database-query module into their own file. Purely internal reorganisation — every existing import keeps working via a re-export; no behaviour change.
